requirements.txt
dateparser==1.1.8
Flask==2.2.3
marshmallow==3.19.0
openai==0.27.8
python-dateutil==2.8.2
python-dotenv==1.0.0
mysql-connector-python==8.0.33
bcrypt==4.0.1
Flask-JWT-Extended==4.5.2